Updated test for signup endpoint
