+++ packages/app/utils/testConstants/constants.js
@@ -1,0 +1,8 @@
+const InitialRequest = {
+  name: "TESTNAME",
+  email: "testname@gmail.com",
+  password: "testname@1234",
+  confirmPassword: "testname@1234",
+};
+
+module.exports = InitialRequest;
